+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ##9.0.1
 * Add support for fa6's 360-degrees icon
 * Fix #154 FaDuotoneIcon explicitly requires IconDataDuotone
+* Use `dart format` instead of deprecated `dartfmt` in the updater 
 
 Thanks to @amkuchta for his work and input