+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-##9.0.1
+##9.1.0
 * Add support for fa6's 360-degrees icon
 * Fix #154 FaDuotoneIcon explicitly requires IconDataDuotone
 * Use `dart format` instead of deprecated `dartfmt` in the updater
 * Automatically enable duotone icons in the example if possible
+* Upgrade to Font Awesome icons 5.15.3
 
 Thanks to @amkuchta for his work and input